ADM-763: [frontend] refactor: extract ErrorMessagePrompt
