+++ frontend/src/components/Common/ReportGrid/ReportCard/style.tsx
@@ -23,19 +23,3 @@ export const StyledReportCardTitle = styled(Typography)({
   fontWeight: 500,
   fontSize: '1rem',
 });
-
-export const StyledErrorSection = styled.div({
-  display: 'flex',
-  flexDirection: 'column',
-  alignItems: 'center',
-  justifyContent: 'center',
-});
-
-export const StyledImgSection = styled.img({
-  height: '3.8rem',
-});
-
-export const StyledErrorMessage = styled.div({
-  color: theme.main.errorMessage.color,
-  fontSize: '0.875rem',
-});
